pyproject.toml
[project]
name = "pyzettel"
version = "0.2.0"
description = "A simple Zettelkasten management system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cattrs>=24.1.2",
    "chardet>=5.2.0",
    "click-loglevel>=0.6.0",
    "click>=8.1.8",
    "click-logging-config>=1.2.0",
    "jinja2>=3.1.5",
    "more-itertools>=10.6.0",
    "platformdirs>=4.3.6",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "types-pyyaml>=6.0.12.20241230",
    "bibtexparser>=1.4.3",
    "rich>=13.9.4",
]

[project.scripts]
pyzettel = "pyzettel.cli.cli:entry_point"
pyzettel-init = "pyzettel.cli.init:init"

[project.optional-dependencies]
ai-rag = [
    "chromadb>=0.6.3",
    "langchain>=0.3.21",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.45",
    "langchain-google-genai>=2.1.1",
    "langchain-openai>=0.3.9",
]
ai = [
    "openai>=1.63.2",
    "numpy>=1.26.2,<=1.26.4",
    "beautifulsoup4>=4.13.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
]
